Update the components
