+++ src/app/iot/places/places.component.html
@@ -7,7 +7,7 @@
             <h2 [title]="location.name">{{location.name | translate }}</h2>
           </div>
         </div>
-        <app-location-row [widgets]="actions.findWidgets(devices, location)" [location]="location"></app-location-row>
+        <app-location-row [widgets]="findWidgets(devices, location)" [location]="location"></app-location-row>
       </div>
     </div>
   </ng-container>
