--- src/app/ng-uikit/ng-uikit.module.ts
@@ -1,38 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { PageContainerComponent } from '@app/ng5-basic/page-container/page-container.component';
-import { ErrorMessageComponent } from '../components/error-message/error-message.component';
-import { TranslateModule, TranslateService } from '@ngx-translate/core';
-import { UiIconBoxComponent } from '@app/uikit/ui-icon-box/ui-icon-box.component';
-import { UiPanelBoxComponent } from '@app/uikit/ui-panel-box/ui-panel-box.component';
-import { UiSmallBoxComponent } from '@app/uikit/ui-small-box/ui-small-box.component';
-import { RouterModule } from '@angular/router';
-import { ProgressLineComponent } from '@app/uikit/progress-line/progress-line.component';
-
-@NgModule({
-  imports: [
-    CommonModule,
-    TranslateModule.forRoot(),
-    RouterModule.forChild([]),
-  ],
-  declarations: [
-    PageContainerComponent,
-    ErrorMessageComponent,
-    UiIconBoxComponent,
-    ProgressLineComponent,
-    UiPanelBoxComponent,
-    UiSmallBoxComponent,
-  ],
-  exports: [
-    PageContainerComponent,
-    ErrorMessageComponent,
-    ProgressLineComponent,
-    UiIconBoxComponent,
-    UiPanelBoxComponent,
-    UiSmallBoxComponent,
-  ],
-  providers: [
-    TranslateService
-  ]
-})
-export class NgUikitModule { }
