+++ src/app/ng5-basic/footer/footer.component.html
