+++ src/app/ng5-basic/progress-line/progress-line.component.ts
