+++ src/app/ng5-basic/ui-panel-box/ui-panel-box.component.scss
