+++ src/app/ng5-basic/ui-small-box/ui-small-box.component.scss
