--- src/app/services/acl.service.ts
@@ -1,8 +1,0 @@
-import { Injectable } from '@angular/core';
-
-@Injectable()
-export class AclService {
-
-  constructor() { }
-
-}
