--- src/app/services/general-interceptor.service.ts
@@ -1,24 +1,0 @@
-import { Injectable } from '@angular/core';
-import {
-  HttpRequest,
-  HttpHandler,
-  HttpEvent,
-  HttpInterceptor,
-} from '@angular/common/http';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/do';
-import 'rxjs/add/operator/catch';
-
-import { ActionsService } from '@app/services/actions.service';
-
-@Injectable()
-export class GeneralInterceptorService implements HttpInterceptor {
-
-  constructor(
-    private actions: ActionsService,
-  ) {}
-
-  intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
-    return next.handle( req );
-  }
-}
