--- src/app/services/globalization.service.spec.ts
@@ -1,15 +1,0 @@
-import { TestBed, inject } from '@angular/core/testing';
-
-import { GlobalizationService } from './globalization.service';
-
-describe('GlobalizationService', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [GlobalizationService]
-    });
-  });
-
-  it('should be created', inject([GlobalizationService], (service: GlobalizationService) => {
-    expect(service).toBeTruthy();
-  }));
-});
