--- src/app/services/notification.service.spec.ts
@@ -1,15 +1,0 @@
-import { TestBed, inject } from '@angular/core/testing';
-
-import { NotificationService } from './notification.service';
-
-describe('NotificationService', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [NotificationService]
-    });
-  });
-
-  it('should be created', inject([NotificationService], (service: NotificationService) => {
-    expect(service).toBeTruthy();
-  }));
-});
